+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Tweaked piece filters UI
   - Clicking on a filter now applies the filter
+  - Name of active filter is now highlighted
 
 ### Fixed